requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_core==0.3.76
langchain_google_genai==2.1.10
python-dotenv==1.1.1
python-multipart==0.0.9
SpeechRecognition==3.10.0
pydub==0.25.1
yt-dlp
